Major update - change config.createDir to Util.CreateDir - move cluster and docker in configuration - change program to follow the new construction of docker and cluster - update frames (main and WorkFlowJInternalFrame) - fix windows toolbox docker connexion - definitively fix issue #1 (by moving the cluster in configuration file) - fix issue #2 (by adding a fake file in src/program/docker and finish to integrate docker-java https://github.com/docker-java/docker-java)

+++ src/configuration/Docker.java
@@ -33,6 +33,7 @@ import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
+import static program.RunProgram.config;
 import com.github.dockerjava.api.DockerClient;
 import com.github.dockerjava.core.*;
 import com.github.dockerjava.api.command.CreateContainerResponse;
@@ -71,9 +72,12 @@ import com.github.dockerjava.core.DockerClientBuilder;
 import com.github.dockerjava.core.command.PullImageResultCallback;
 import com.github.dockerjava.core.util.CompressArchiveUtil;
 import com.github.dockerjava.core.command.ExecStartResultCallback;
+import java.io.BufferedWriter;
 import java.io.ByteArrayOutputStream;
+import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.io.OutputStreamWriter;
 import java.util.concurrent.TimeUnit;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -87,7 +91,7 @@ public class Docker {
     public static final SimpleDateFormat dateyear = new SimpleDateFormat("yyyy");
     //--Logging function
     public PrintWriter log;
-    public static String kword = "armadilloWF";
+    public static String kword = "armadillo_WF_";
     
     public Docker() {}
     
@@ -109,9 +113,14 @@ public class Docker {
      * Functions to test if docker is present
      */
     public static boolean isDockerHere () {
-        if (isDockerHereByRootDir() || isDockerHereByVersion())
-            return true;
-        return false;
+        boolean b = false;
+        try{
+            b = isDockerHereByVersion();
+        } catch (Exception e) { 
+            System.out.println(Util.BRDockerNotFound());
+            b = false;
+        } 
+        return b;
     }
     
     public static boolean isDockerHereByRootDir () {
@@ -123,11 +132,18 @@ public class Docker {
     }
 
     public static boolean isDockerHereByVersion () {
-        Version v = getVersion();
-        if (v.getGoVersion().length() > 0)
-            return true;
-        if (v.getVersion().length() > 0)
-            return true;
+        Version v;
+        try {
+            v = getVersion();
+        } catch (Exception e) { 
+            v = null;
+        }
+        if (v!=null){
+            if (v.getGoVersion().length() > 0)
+                return true;
+            if (v.getVersion().length() > 0)
+                return true;
+        }
         return false;
     }
     
@@ -136,14 +152,68 @@ public class Docker {
      * COULD BE GOOD to create a similar function with specific config builder for advanced docker user
      */
 
-    public static DockerClientConfig createDockerClientConfig () {
-        return DefaultDockerClientConfig.createDefaultConfigBuilder()
-                .build();
+    public static boolean isDockerNeededInfoInConfig() {
+        List<String> lines = Arrays.asList("DockerHost","DockerTlsVerify","DockerCertPath","DockerConfig");
+        return Util.isListInProperties(config.properties,lines);
     }
     
-    public static DockerClient getDockerClient () {
-        return DockerClientBuilder.getInstance(createDockerClientConfig())
+    public static DockerClientConfig getDockerClientConfigWindowsSimple () {
+        DockerClientConfig dcc = null;
+        if (isDockerNeededInfoInConfig()){
+            return null;
+        } else {
+            try { 
+                dcc = DefaultDockerClientConfig.createDefaultConfigBuilder()
+                    .withDockerHost("tcp://"+config.get("DockerHost"))
+                    .withDockerTlsVerify(config.get("DockerTlsVerify"))
+                    .withDockerCertPath(config.get("DockerCertPath"))
+                    .withDockerConfig(config.get("DockerConfig"))
+                    .build();
+            } catch (Exception e) { 
+                System.out.println(Util.BRDockerAPINotAccessible());
+                dcc = null;
+            } 
+        }
+        return dcc;
+    }
+    
+    public static DockerClientConfig getDockerClientConfigUnix () {
+        DockerClientConfig dcc = null;
+        try {
+            dcc = DefaultDockerClientConfig.createDefaultConfigBuilder()
                 .build();
+        } catch (Exception e) { 
+            System.out.println(Util.BRDockerAPINotAccessible());
+            dcc = null;
+        } 
+        return dcc;
+    }
+    
+    public static DockerClientConfig getDockerClientConfig(){
+        DockerClientConfig dcc = null;
+        if (dcc == null)
+            if (config.getBoolean("Windows"))
+                return getDockerClientConfigWindowsSimple();
+            else
+                return getDockerClientConfigUnix();
+        return dcc;
+    }
+
+    
+    public static DockerClient getDockerClient () {
+        DockerClient dc;
+        try {
+            DockerClientConfig dcc = getDockerClientConfig();
+            if (dcc!=null){
+                dc = DockerClientBuilder.getInstance(dcc).build();
+            }else{
+                dc = null;
+            }
+        } catch (Exception e) { 
+            System.out.println(Util.BRDockerUnableToCreateClient());
+            dc = null;
+        } 
+        return dc;
     }    
     
     public static boolean closeDockerClient (DockerClient dockerClient) {
@@ -156,15 +226,39 @@ public class Docker {
         }
         return false;
     }
-
+    
+    public static DockerClientConfig createDockerClientAdvancedConfig (workflow_properties properties) {
+        String DockerHost = properties.get("DockerHost_Text");
+        String DockerTlsVerify = properties.get("DockerTlsVerify_Text");
+        String DockerCertPath = properties.get("DockerCertPath");
+        String DockerRegistryUserName = properties.get("DockerRegistryUserName_Text");
+        String DockerRegistryPassword = properties.get("DockerRegistryPassword_Text");
+        String DockerRegistryEmail = properties.get("DockerRegistryEmail_Text");
+        String DockerRegistryUrl = properties.get("DockerRegistryUrl_Text");
+         
+        DockerClientConfig config = DefaultDockerClientConfig.createDefaultConfigBuilder()
+            .withDockerHost(DockerHost)
+            .withDockerTlsVerify(DockerTlsVerify)
+            .withDockerCertPath(DockerCertPath)
+            .withRegistryUsername(DockerRegistryUserName)
+            .withRegistryPassword(DockerRegistryPassword)
+            .withRegistryEmail(DockerRegistryEmail)
+            .withRegistryUrl(DockerRegistryUrl)
+            .build();
+        return config;
+    }
+    
     /*
      * Functions to get info from Docker
      */    
 
     public static Version getVersion (){
         DockerClient dockerClient = getDockerClient();
-        Version version = dockerClient.versionCmd().exec();
-        closeDockerClient(dockerClient);
+        Version version = null;
+        if (dockerClient!=null){
+            version = dockerClient.versionCmd().exec();
+            closeDockerClient(dockerClient);
+        }
         return version;
     }
     
@@ -405,7 +499,7 @@ public class Docker {
             String   id = containers.get(i).getId();
             String[] names = containers.get(i).getNames();
             for (String name : names){
-                if (name.toLowerCase().contains("armadillo_wf_")){
+                if (name.toLowerCase().contains(kword)){
                     list.add(containers.get(i));
                 }
             }
@@ -428,7 +522,7 @@ public class Docker {
         for(int i=0; i < containers.size(); i++){
             String[] names = containers.get(i).getNames();
             for (String name : names){
-                if (name.toLowerCase().contains("armadillo_wf_")){
+                if (name.toLowerCase().contains(kword)){
                     list.add(name);
                 }
             }
@@ -472,7 +566,7 @@ public class Docker {
             String   id = containers.get(i).getId();
             String[] names = containers.get(i).getNames();
             for (String name : names){
-                if (name.toLowerCase().contains("armadillo_wf_")){
+                if (name.toLowerCase().contains(kword)){
                     list.add(id);
                 }
             }
@@ -491,7 +585,7 @@ public class Docker {
             String   id = containers.get(i).getId();
             String[] names = containers.get(i).getNames();
             for (String name : names){
-                if (name.toLowerCase().contains("armadillo_wf_")){
+                if (name.toLowerCase().contains(kword)){
                     list.add(id);
                 }
             }
@@ -561,7 +655,7 @@ public class Docker {
         if(addImage(dockerImage)){
             String alea = Integer.toString(new SecureRandom().nextInt());
             alea = alea.replaceAll("-","");
-            String containerName = "armadillo_wf_"+alea;
+            String containerName = kword+alea;
             DockerClient dockerClient = getDockerClient();
             Volume[] list_v = createListVolumes(sharedFolders);
             Bind[] list_bd = createListBinds(sharedFolders,list_v);
@@ -595,7 +689,6 @@ public class Docker {
         }
         return false;
     }
-    
     
     /**
      * Start containers functions
@@ -643,9 +736,11 @@ public class Docker {
      * Remove containers functions
      */
     public static long removeContainer(workflow_properties properties) {
-        String DockerContainerID = properties.get("DockerContainerID");
         long startTime = System.nanoTime();
-        removeContainer(DockerContainerID);
+        if (properties.isSet("DockerContainerID")){
+            String DockerContainerID = properties.get("DockerContainerID");
+            //removeContainer(DockerContainerID);
+        }
         long endTime = System.nanoTime();
         long duration = (endTime - startTime);
         return TimeUnit.SECONDS.convert(duration, TimeUnit.NANOSECONDS);
@@ -729,6 +824,7 @@ public class Docker {
      * Execute the bash file
      */
     public static void executeBashFile(workflow_properties properties) {
+        String[] cmd = {"/bin/sh","-c","/data/dockerBash.sh"};
         ByteArrayOutputStream stdout = new ByteArrayOutputStream();
         ByteArrayOutputStream stderr = new ByteArrayOutputStream();
         String DockerContainerID = properties.get("DockerContainerID");
@@ -737,12 +833,13 @@ public class Docker {
         ExecCreateCmdResponse execCreateCmdResponse = dockerClient.execCreateCmd(DockerContainerID)
                 .withAttachStdout(true)
                 .withAttachStderr(true)
-                .withCmd("/data/dockerBash.sh", "/data/dockerBash.log")
+                .withCmd(cmd)
                 .withUser("root")
                 .exec();
         try {
-            dockerClient.execStartCmd(execCreateCmdResponse.getId()).exec(
-                    new ExecStartResultCallback(stdout, stderr)).awaitCompletion();
+            dockerClient.execStartCmd(execCreateCmdResponse.getId())
+                    .exec(new ExecStartResultCallback(stdout, stderr))
+                    .awaitCompletion();
             String dockerSTD = stdout.toString()+stderr.toString();
             dockerSTD = dockerSTD.replaceAll("groupadd: group '\\w*' already exists", "");
             properties.put("DockerSTD",dockerSTD);
@@ -757,41 +854,59 @@ public class Docker {
      */
     public static long prepareDockerBashFile(workflow_properties properties, String dockerCli) {
         long startTime = System.nanoTime();
-        String s = Util.currentPath()+File.separator+"tmp"+File.separator+"dockerBash.sh";
-        Util.CreateFile(s);
-        Path file = Paths.get(s);
-        
-        try {
-            file.toFile().setExecutable(true, false);
-        }  catch(Exception ex) {
-            System.out.println("Execute permission on bash file is not set!");
-            System.out.println(ex);
+        if (Docker.isDockerHere()){
+            String userName = "";
+            String grpName = "";
+            String dockerAddUser = "";
+            String dockerAddGrp = "";
+            String dockerChangeOwner = "";
+            String dockerBash = "";
+            String exitValue = "echo exit is -$?-";
+            String timeStart= "";
+            String timeEnd  = "";
+            String timeDiff = "";
+            String timeEcho = "";
+            if (!config.getBoolean("Windows")){
+                userName = Util.getOwnerJar();
+                grpName = Util.getGroupJar();
+                dockerAddGrp = "groupadd "+grpName+"";
+                dockerAddUser = "useradd "+userName+"";
+                dockerChangeOwner = "chown -R "+userName+":"+grpName+" /data/outputs/";
+                dockerBash = "#!/bin/bash";
+                exitValue = "echo exit is -$?-";
+                timeStart= "START=$(date +%s)";
+                timeEnd  = "END=$(date +%s)";
+                timeDiff = "DIFF=$(( $END - $START ))";
+                timeEcho = "echo \"It took $DIFF seconds\"";
+            } else {
+                dockerBash = "#/bin/bash";
+            }
+
+            String eol = "\n";
+            String cli = dockerBash+eol+timeStart+eol+dockerCli+eol+exitValue+eol+timeEnd+eol+timeDiff+eol+timeEcho+eol+dockerAddUser+eol+dockerAddGrp+eol+dockerChangeOwner;
+
+            String s = Util.currentPath()+File.separator+"tmp"+File.separator+"dockerBash.sh";
+            Util.CreateFile(s);
+            Path path = Paths.get(s);
+            try {
+                path.toFile().setExecutable(true, false);
+            }  catch(Exception ex) {
+                System.out.println("Execute permission on bash file is not set!");
+                System.out.println(ex);
+            }
+            try {
+                File file = path.toFile();
+                BufferedWriter out = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(file),"UTF-8"));
+                out.write(cli);
+                out.close();
+            } catch (IOException e) {
+                System.out.println("Exception "+e);
+            }
+            String DockerContainerID = properties.get("DockerContainerID");
+            properties.put("DockerCommandLine",dockerCli);
+            copyFileToContainer(DockerContainerID,s,"/data/");
+            Util.deleteFile(s);
         }
-        
-        String userName = Util.getOwnerJar();
-        String grpName = Util.getGroupJar();
-        String dockerAddUser = "useradd "+userName+"";
-        String dockerAddGrp = "groupadd "+grpName+"";
-        String dockerChangeOwner = "chown -R "+userName+":"+grpName+" /data/outputs/";
-        String exitValue = "echo exit is -$?-";
-        String timeStart= "START=$(date +%s)";
-        String timeEnd  = "END=$(date +%s)";
-        String timeDiff = "DIFF=$(( $END - $START ))";
-        String timeEcho = "echo \"It took $DIFF seconds\"";
-        
-        List<String> lines = Arrays.asList("#!/bin/bash", timeStart, dockerCli, exitValue, timeEnd, timeDiff, timeEcho,dockerAddUser, dockerAddGrp, dockerChangeOwner);
-        try {
-            Files.write(file, lines);
-        } catch (IOException ex) {
-            System.out.println("Unable to create the bash file in the container");
-            System.out.println(ex);
-        }
-        
-        String DockerContainerID = properties.get("DockerContainerID");
-        properties.put("DockerCommandLine",dockerCli);
-        copyFileToContainer(DockerContainerID,s,"/data/");
-        Util.deleteFile(s);
-        
         long endTime = System.nanoTime();
         long duration = (endTime - startTime);
         return TimeUnit.SECONDS.convert(duration, TimeUnit.NANOSECONDS);
@@ -820,12 +935,28 @@ public class Docker {
         HashMap<String,String> sharedFolders = new HashMap<String,String>();
         for(int i = 0; i < tabPath.length; i++) {
             if (tabPath[i]!="") {
-                String dir = Util.getParentOfFile(tabPath[i]);
-                dir = Util.getCanonicalPath(dir);
-                sharedFolders.put(dir, doInputs+tabID[i]+"/");
+                sharedFolders = addInSharedFolder(sharedFolders,tabPath[i],doInputs+tabID[i]+"/");
             }
         }
         return sharedFolders;
+    }
+    
+    public static HashMap<String,String> addInSharedFolder(HashMap<String,String> sf, String outputPath, String doOutputs) {
+        String dir = Util.getCanonicalPath(outputPath);
+        dir = Util.getParentOfFile(dir);
+        dir = Util.getCanonicalPath(dir);
+        dir = cleanWindowsPath4DockerVolume(dir);
+        sf.put(dir, doOutputs);
+        return sf;
+    }
+    
+    public static String cleanWindowsPath4DockerVolume(String path){
+        if (config.getBoolean("Windows")){
+            path = path.substring(0, 1).toLowerCase() + path.substring(1);
+            path = path.replaceAll("\\\\", "/");
+            path = path.replaceFirst("(\\w):","/$1".toLowerCase());
+        }
+        return path;
     }
     
     /**
